linked activity to CourseInformation
