+++ app/src/main/java/com/example/al4t_claco/controller/Dashboard.kt
@@ -15,6 +15,7 @@ import com.example.al4t_claco.LoginActivity
 import com.example.al4t_claco.R
 import com.example.al4t_claco.model.Activity
 import com.example.al4t_claco.model.Course
+import com.example.al4t_claco.model.File
 import com.google.android.material.navigation.NavigationView
 
  class Dashboard  : AppCompatActivity() {
@@ -49,15 +50,24 @@ import com.google.android.material.navigation.NavigationView
             }
             false
         }
+
+        //TODO : Bouger ca autre part mais pas ultra important pour le moment
+
         val activity = Activity("Info","4inf", listOf("Lorge","Lurkin","Dekimpe"))
         val activity2 = Activity("Bidule","4inf", listOf("Lorge","Lurkin","Dekimpe"))
-        val activityList = listOf<Activity>(activity,activity2)
+        val activity3 = Activity("Truc","4inf", listOf("Lorge","Lurkin","Dekimpe"))
+
+        activity.resources = listOf(File("file 1","pdf"), File("file 2","pdf"), File("file 3","pdf"), File("file 4","PDF"))
+        activity2.resources = listOf(File("file 1","pdf"), File("file 2","pdf"), File("file 3","pdf"), File("file 4","PDF"))
+        activity3.resources = listOf(File("file 5","pdf"), File("file 6","pdf"), File("file 3","pdf"), File("file 4","PDF"))
+
+        val activityList = listOf<Activity>(activity,activity2,activity3)
 
         val course = Course("Info","4inf",4,5,"Lur", activityList)
         val course1 = Course("Database","4DB",4,5,"Lor", activityList)
         val course2 = Course("APPS","4app",4,5,"LRK", activityList)
         val course3 = Course("Electronics","4el",4,5,"MCH", activityList)
-        val course4 = Course("Electicity","4inf",4,5,"CMS", activityList)
+        val course4 = Course("AAAAAAA","4inf",4,5,"CMS", activityList)
 
         val courseNames = listOf<Course>(course,course1,course2,course3,course4);
